refactor question list to have a parent fetching and passing data for it
