+++ js/components/home/home.jsx
@@ -1,12 +1,11 @@
 import Header from "./header.jsx";
-import QuestionList from "./question_list.jsx";
+import QuestionPanel from "./question_panel.jsx";
 
 export default function Home() {
   return (
     <div>
       <Header />
-      {/* TODO filter & paginate pages, i.e. p1, 2, 3 */}
-      <QuestionList />
+      <QuestionPanel />
     </div>
   );
 }
