+++ js/components/home/question_list.jsx
@@ -1,51 +1,17 @@
-import { useEffect, useState } from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import {
-  faCheck, faHourglassEnd, faClock, faMinus,
-  faAnglesLeft, faAngleLeft, faAngleRight, faAnglesRight
-} from "@fortawesome/free-solid-svg-icons";
+import { faCheck, faHourglassEnd, faClock, faMinus } from "@fortawesome/free-solid-svg-icons";
 import { Tooltip } from "react-tooltip";
-import { getReqHeader, capitalizeFirst } from "../../utils/utils.js";
+import { capitalizeFirst } from "../../utils/utils.js";
 
-export default function QuestionList() {
-  const [questions, setQuestions] = useState([]);
-  const [page, setPage] = useState(1);
-  const [totalPages, setTotalPages] = useState(1);
-
-  useEffect(async () => {
-    const data = await getQuestions();
-    setQuestions(data.data);
-    setPage(data.page);
-    setTotalPages(data.pages);
-  }, []);
-
-  const getQuestions = async (queries) => {
-    const req = {
-      method: "GET",
-      header: getReqHeader(),
-    }
-
-    return await fetch(`/api/questions?${queries ? queries : ""}`, req)
-      .then(res => res.json())
-      .then(json => json.data)
-      .catch(err => {
-        throw err;
-      })
-  };
-
-  const fetchForPage = (page) => {
-
-  };
-
+export default function QuestionList(props) {
   return (
     <div className="mt-4 text-sm w-[100%] flex flex-col justify-start items-center">
-      {questions.map((q, i) => {
+      {props.questions.map((q, i) => {
         return (<ListItem key={"q" + i}
           idx={i + 1} title={q.title} status={q.status} source={q.source}
           difficulty={q.difficulty} timestamp={q.last_modified} tags={q.tags}
         />);
       })}
-      <PaginationBoxes page={page} totalPages={totalPages} fetchForPage={fetchForPage} />
     </div>
   );
 }
@@ -108,55 +74,6 @@ function ListItem(props) {
             </div>);
         })}
       </div>
-    </div>
-  );
-}
-
-function PaginationBoxes(props) {
-  const [boxes, setBoxes] = useState([]);
-
-  useEffect(() => {
-    // Maximum of 9 boxes, should be an odd number for style
-    const numBoxes = Math.min(props.totalPages, 9);
-    let front = [];
-    let back = [];
-    let res = [];
-
-    for (let i = 0; i < numBoxes; i++) {
-      if (front.length + back.length + 1 < numBoxes) {
-        if (props.page - i - 1 >= 1) {
-          front.unshift(props.page - i - 1);
-        }
-
-        if (props.page + i + 1 <= props.totalPages) {
-          back.push(props.page + i + 1);
-        }
-      }
-    }
-
-    res.push(...front, props.page, ...back);
-    setBoxes(res);
-  }, [props.page, props.totalPages]);
-
-  return (
-    <div>
-      <Box page={1} icon={faAnglesLeft} />
-      <Box page={Math.max(props.page - 1, 1)} icon={faAngleLeft} />
-      {boxes.map(boxNum => {
-        return <Box page={boxNum} />;
-      })}
-      <Box page={Math.min(props.page + 1, props.totalPages)} icon={faAngleRight} />
-      <Box page={props.totalPages} icon={faAnglesRight} />
-    </div>
-  );
-}
-
-function Box(props) {
-  return (
-    <div>
-      {props.icon
-        ? <FontAwesomeIcon icon={props.icon} />
-        : props.page}
     </div>
   );
 }
